:sparkles: migrate legend
