+++ src/Spillgebees.Blazor.Map.Docs/Pages/Components/MapLegendPage.razor
@@ -11,7 +11,8 @@
 <p>An interactive, toggleable legend control that wires layer visibility groups to a customizable item template.</p>
 
 <h2 id="overview">// overview</h2>
-<p>Place <code>MapLegend</code> as a child of <code>SgbMap</code>. It renders a collapsible control panel whose items map to layer visibility groups on the map. When the user toggles an item, the corresponding layers show or hide.</p>
+<p>Configure legend content directly on <code>LegendMapControl</code> inside <code>SgbMap.Controls</code>. This control-driven contract renders a collapsible panel whose items map to layer visibility groups on the map.</p>
+<p><code>MapLegend</code> is now a legacy migration component and throws at runtime. This is an intentional breaking change in Phase 1 so consumers move to <code>LegendMapControl.Content</code>.</p>
 
 <h2 id="definition">// definition</h2>
 <p>Build a <code>MapLegendDefinition</code> to describe the legend structure. Each section contains items with an id, label, optional description, and a list of layer IDs to control.</p>
@@ -20,38 +21,48 @@
 {
     ["Legend definition"] =
 @"var legend = new MapLegendDefinition(
-    title: ""Map layers"",
-    sections:
+    Sections:
     [
-        new MapLegendSection(""Infrastructure"",
-        [
-            new MapLegendItem(
-                id: ""stations"",
-                label: ""Stations"",
-                description: ""Railway stations"",
-                layerIds: [""stations-circle"", ""stations-label""]),
-            new MapLegendItem(
-                id: ""tracks"",
-                label: ""Tracks"",
-                layerIds: [""tracks-line""]),
-        ]),
+        new MapLegendSectionDefinition(
+            ""Infrastructure"",
+            [
+                new MapLegendItemDefinition(
+                    ""stations"",
+                    ""Stations"",
+                    ""Railway stations"",
+                    [new MapLegendTargetDefinition(""base-style"", [""stations-circle"", ""stations-label""])],
+                    IsVisibleByDefault: true,
+                    IsToggleable: true
+                ),
+                new MapLegendItemDefinition(
+                    ""tracks"",
+                    ""Tracks"",
+                    Targets: [new MapLegendTargetDefinition(""base-style"", [""tracks-line""])],
+                    IsVisibleByDefault: true,
+                    IsToggleable: true
+                ),
+            ]
+        ),
     ]);",
 })" />
 
 <h2 id="control_options">// control_options</h2>
-<p><code>MapLegend</code> binds to a legend control entry declared in <code>SgbMap.Controls</code> via <code>ControlId</code>. Position and ordering are defined exclusively on that map control entry.</p>
+<p><code>LegendMapControl</code> now carries placement, chrome, and content in constructor-complete grouped options.</p>
 
 <CodeBlock Language="csharp" Files="@(new Dictionary<string, string>
 {
     ["Control options"] =
 @"var controls = new List<MapControl>
 {
-    new LegendMapControl(""overlay-legend"", Position: ControlPosition.TopRight, Collapsible: true),
+    new LegendMapControl(
+        ""overlay-legend"",
+        new MapControlPlacement(ControlPosition.TopRight, 500, Enabled: true),
+        new LegendChromeOptions(""Legend"", Collapsible: true, InitiallyOpen: true, ClassName: null),
+        new LegendContentOptions(legend, ItemTemplate: null, OnItemVisibilityChanged: default)
+    ),
 };
 
-<SgbMap Controls=""@controls"">
-    <MapLegend ControlId=""overlay-legend"" Definition=""@legend"" />
-</SgbMap>",
+<SgbMap Controls=""@controls"" />",
 })" />
 
 <h2 id="custom_templates">// custom_templates</h2>
