+++ src/Spillgebees.Blazor.Map.Docs/Pages/Components/SgbMapPage.razor
@@ -23,7 +23,7 @@
 </ExampleView>
 
 <h2 id="custom_controls">// custom_controls</h2>
-<p>Custom control components (like <code>MapLegend</code>) bind to a declarative control shell in <code>SgbMap.Controls</code> by matching <code>ControlId</code>. The shell controls map position and order, and the child component only provides content and behavior.</p>
+<p>Custom controls are configured directly in <code>SgbMap.Controls</code>. For legends, use <code>LegendMapControl</code> with grouped content options to define legend sections and items, while the control shell sets map position and order.</p>
 <p>Ordering is deterministic per map corner: controls are sorted by <code>Order</code> (ascending), then declaration order in <code>Controls</code>, then <code>ControlId</code> as a stable tie-breaker. Keep each <code>ControlId</code> unique and pass the same id to the matching child control component.</p>
 <ExampleView TComponent="CustomControlsExample">
     <CustomControlsExample />
