+++ src/Spillgebees.Blazor.Map.Docs/Pages/Components/TrackedDataSourcePage.razor
@@ -23,7 +23,7 @@
     <li>Smooth position interpolation via <code>AnimationOptions</code></li>
     <li>SVG-to-data-URI custom icon registration via the map <code>Images</code> parameter</li>
     <li>Style composition with a railway overlay on a Positron base</li>
-    <li>Interactive <a href="components/map-legend">MapLegend</a> with custom swatch templates</li>
+    <li>Interactive <a href="components/map-legend">LegendMapControl</a> configured in <code>SgbMap.Controls</code>, with custom swatch templates</li>
 </ul>
 
 <ExampleView TComponent="TrainTrackingExample"
