+++ src/Spillgebees.Blazor.Map.Tests/BaseMapControlsTests.cs
@@ -1,6 +1,7 @@
 using AwesomeAssertions;
 using Spillgebees.Blazor.Map.Components;
 using Spillgebees.Blazor.Map.Models.Controls;
+using Spillgebees.Blazor.Map.Models.Legends;
 
 namespace Spillgebees.Blazor.Map.Tests;
 
@@ -11,7 +12,12 @@ public class BaseMapControlsTests
     {
         // arrange
         var map = new TestBaseMap();
-        var control = new LegendMapControl("legend-main");
+        var control = new LegendMapControl(
+            "legend-main",
+            new MapControlPlacement(ControlPosition.TopRight, 500, Enabled: true),
+            new LegendChromeOptions("Legend", Collapsible: true, InitiallyOpen: true, ClassName: null),
+            new LegendContentOptions(new MapLegendDefinition([]), ItemTemplate: null, OnItemVisibilityChanged: default)
+        );
         map.SetInternalControls([control]);
 
         // act
@@ -28,7 +34,18 @@ public class BaseMapControlsTests
     {
         // arrange
         var map = new TestBaseMap();
-        map.SetInternalControls([new LegendMapControl("legend-main")]);
+        map.SetInternalControls([
+            new LegendMapControl(
+                "legend-main",
+                new MapControlPlacement(ControlPosition.TopRight, 500, Enabled: true),
+                new LegendChromeOptions("Legend", Collapsible: true, InitiallyOpen: true, ClassName: null),
+                new LegendContentOptions(
+                    new MapLegendDefinition([]),
+                    ItemTemplate: null,
+                    OnItemVisibilityChanged: default
+                )
+            ),
+        ]);
 
         // act
         var found = map.TryResolveControl("missing", out var resolvedControl);
