+++ src/Spillgebees.Blazor.Map/Models/Legends/MapLegendVisibilityChangedEventArgs.cs
@@ -1,8 +1,8 @@
 namespace Spillgebees.Blazor.Map.Models.Legends;
 
 /// <summary>
-/// Event arguments raised when a legend item changes visibility.
+/// Event arguments raised when a legend item changes selection.
 /// </summary>
 /// <param name="Item">The toggled legend item.</param>
-/// <param name="Visible">The new visibility value.</param>
-public sealed record MapLegendVisibilityChangedEventArgs(MapLegendItemDefinition Item, bool Visible);
+/// <param name="Selected">The new selection value.</param>
+public sealed record MapLegendVisibilityChangedEventArgs(MapLegendItemDefinition Item, bool Selected);
